styles: Update hero styles and layout measures
The commit updates the hero and its children styles (Heading, Call to actions, tagline) and it adds layout measures to control the maximum width of the body to prevent stretching of content in larger screens. It also removes unnecessary code and improves the responsiveness of the navigation and hero sections.

Finally, we add the functionality of looking for the preferred color scheme of the user to set the page theme